feat : 데미지 처리 및 스킬 캐스팅레벨 조절 UI 추가

+++ URPRagProjectSingle/Assets/Characters/Skills/Scripts/DataReader.cs
@@ -205,7 +205,7 @@ public class SkillBaseObjectOnly
     public ValueType damageType;
     public float coefficient;
     public ValueType coefficientType;
-    public float skillBound;
+    public byte skillBound;
     public float spCost;
     public float defaultCastingTime;
     public byte skillRange;
